Marshalling return value correctly to check for errors

+++ ffi.cs
@@ -7,7 +7,7 @@ namespace NTRUMLS.ffi {
     public static class ffi {
 
         [DllImport("ntrumls")]
-        public static extern IntPtr pq_gen_key(ParamSet p, out IntPtr privkey_blob_len, out byte[] privkey_blob, out IntPtr pubkey_blob_len, out byte[] pubkey_blob);
+        public static extern int pq_gen_key(ParamSet p, out IntPtr privkey_blob_len, out byte[] privkey_blob, out IntPtr pubkey_blob_len, out byte[] pubkey_blob);
 
         [DllImport("ntrumls")]
         public static extern IntPtr pq_sign(IntPtr packed_sig_len, byte[] packed_sig, IntPtr private_key_len, byte[] private_key_blob, IntPtr public_key_len, byte[] public_key_blob, IntPtr msg_len, byte[] msg);
